pyproject.toml
[project]
name = "telegram-feed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "click>=8.3.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "telethon>=1.42.0",
    "tomli-w>=1.2.0",
]
